Attempt to fix paths in test yamls, fixes #4
Analysis:
If the python test is not run from the catheter_utils directory,
the relative paths referred to in the test yaml settings for
the test_localization code will fail to resolve and the test will fail.

Resolution:
Attempting to fix up non-resolving paths by prepending the path
with the catheter_utils directory. Only the testdata_path &
target_path are checked.

The existing validate_settings_yaml method is still used to
validate the path and will exit the test if the path still does
not resolve.