Adding snr values to projections for #2
Analysis:
SNR values were not being computed with projections. These are useful
for analysing catheter tracking results.

Resolution:
Computing SNR values for projections on initialization of FindData.
These are organized by coil and axis, with SNRs for each readout. A separate
SNR array can be computed per coil, combining the axes as specified by
the caller (min / median / mean).

The SNR calculation uses the size of the peak over the standard
deviation of the noise region of the projection (this method was already
in the projections module, but not being called).

Combination of snr's is now done in "combine_snr" and the
combinaton method is a property, set to 'min' by default.

The combine_snr method is called after the projections
are read & if the combination method is changed.

Docstrings have been updated