Returning iterations required for localization #6
Analysis:
Need to return the number of iterations for algorithms
other than JPNG.

Resolution:
The PNG algorithm (iterative weighted centroid) now
returns the number of iterations used.

The localize_coil and localize_catheter functions
now also return the iterations (zero for non-iterative
algorithms):
- localize_coil returns the iterations required for
each projection in an array.
- localize_catheter, which operates across two coils
and 3 or 4 projections per coil, returns the largest
number of iterations required

Additionally, logging has been updated to reflect
the actual number of iterations (it was off by
one).

Note that wjpng is not fully tested: it errors out
of some inputs