Returning centroid with distances from centroid
Analysis:
get_distances_from_centroid calls the appropriate
get_centroid function and returns distances from
the centroid. It would be useful to have it
return the centroid as well, to avoid recomputing
this in another call.

Resolution:
get_distances_from_centroid now returns the centroid
as well as the distances from it.

Testing should be updated to check that the returned
centroid is correct.